requirements.txt
aiohttp==3.7.4.post0
astroid==2.11.7
async-timeout==3.0.1
attrs==22.1.0
autopep8==1.6.0
chardet==4.0.0
dill==0.3.5.1
discord.py==1.7.3
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
multidict==6.0.2
packaging==21.3
platformdirs==2.5.2
pycodestyle==2.9.1
pylint==2.14.5
pyparsing==3.0.9
python-dotenv==0.20.0
pytoolconfig==1.2.2
rope==1.3.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
typing_extensions==4.3.0
wrapt==1.14.1
yarl==1.8.1
